+++ tests/IndicesTest.php
@@ -4,15 +4,16 @@ namespace MarketDataApp\Tests;
 
 use Carbon\Carbon;
 use GuzzleHttp\Client as GuzzleClient;
+use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use MarketDataApp\Client;
-use MarketDataApp\Endpoints\Responses\IndicesCandle;
-use MarketDataApp\Endpoints\Responses\IndicesCandles;
-use MarketDataApp\Endpoints\Responses\IndicesQuote;
+use MarketDataApp\Endpoints\Responses\Indices\Candle;
+use MarketDataApp\Endpoints\Responses\Indices\Candles;
+use MarketDataApp\Endpoints\Responses\Indices\Quote;
 use PHPUnit\Framework\TestCase;
 
 class IndicesTest extends TestCase
@@ -27,7 +28,7 @@ class IndicesTest extends TestCase
         $this->client = $client;
     }
 
-    public function testIndicesQuote_success()
+    public function testQuote_success()
     {
         $mocked_response = [
             's'          => 'ok',
@@ -42,7 +43,7 @@ class IndicesTest extends TestCase
         $this->setMockResponses([new Response(200, [], json_encode($mocked_response))]);
 
         $response = $this->client->indices->quote("DJI");
-        $this->assertInstanceOf(IndicesQuote::class, $response);
+        $this->assertInstanceOf(Quote::class, $response);
         $this->assertEquals($mocked_response['s'], $response->status);
         $this->assertEquals($mocked_response['symbol'][0], $response->symbol);
         $this->assertEquals($mocked_response['last'][0], $response->last);
@@ -53,7 +54,29 @@ class IndicesTest extends TestCase
         $this->assertEquals(Carbon::parse($mocked_response['updated']), $response->updated);
     }
 
-    public function testIndicesCandles_fromTo_success()
+    public function testQuote_noData_success()
+    {
+        $mocked_response = [
+            's' => 'no_data',
+        ];
+        $this->setMockResponses([new Response(200, [], json_encode($mocked_response))]);
+
+        $response = $this->client->indices->quote("DJI");
+        $this->assertInstanceOf(Quote::class, $response);
+        $this->assertEquals($mocked_response['s'], $response->status);
+        $this->assertFalse(isset($response->symbol));
+        $this->assertFalse(isset($response->last));
+        $this->assertFalse(isset($response->change));
+        $this->assertFalse(isset($response->change_percent));
+        $this->assertFalse(isset($response->fifty_two_week_high));
+        $this->assertFalse(isset($response->fifty_two_week_low));
+        $this->assertFalse(isset($response->updated));
+    }
+
+    /**
+     * @throws GuzzleException
+     */
+    public function testCandles_fromTo_success()
     {
         $mocked_response = [
             's' => 'ok',
@@ -73,18 +96,67 @@ class IndicesTest extends TestCase
         );
 
         // Verify that the response is an object of the correct type.
-        $this->assertInstanceOf(IndicesCandles::class, $response);
+        $this->assertInstanceOf(Candles::class, $response);
         $this->assertCount(5, $response->candles);
 
         // Verify each item in the response is an object of the correct type and has the correct values.
-        for($i = 0; $i < count($response->candles); $i++) {
-            $this->assertInstanceOf(IndicesCandle::class, $response->candles[$i]);
+        for ($i = 0; $i < count($response->candles); $i++) {
+            $this->assertInstanceOf(Candle::class, $response->candles[$i]);
             $this->assertEquals($mocked_response['c'][$i], $response->candles[$i]->close);
             $this->assertEquals($mocked_response['h'][$i], $response->candles[$i]->high);
             $this->assertEquals($mocked_response['l'][$i], $response->candles[$i]->low);
             $this->assertEquals($mocked_response['o'][$i], $response->candles[$i]->open);
             $this->assertEquals(Carbon::parse($mocked_response['t'][$i]), $response->candles[$i]->timestamp);
         }
+    }
+
+    /**
+     * @throws GuzzleException
+     */
+    public function testCandles_noData_success()
+    {
+        $mocked_response = [
+            's' => 'no_data',
+        ];
+        $this->setMockResponses([new Response(200, [], json_encode($mocked_response))]);
+
+        $response = $this->client->indices->candles(
+            symbol: "DJI",
+            from: Carbon::parse('2022-09-01'),
+            to: Carbon::parse('2022-09-05'),
+            resolution: 'D'
+        );
+
+        // Verify that the response is an object of the correct type.
+        $this->assertInstanceOf(Candles::class, $response);
+        $this->assertEquals($mocked_response['s'], $response->status);
+        $this->assertEmpty($response->candles);
+        $this->assertFalse(isset($response->next_time));
+    }
+
+    /**
+     * @throws GuzzleException
+     */
+    public function testCandles_noDataNextTime_success()
+    {
+        $mocked_response = [
+            's'        => 'no_data',
+            'nextTime' => 1659326400,
+        ];
+        $this->setMockResponses([new Response(200, [], json_encode($mocked_response))]);
+
+        $response = $this->client->indices->candles(
+            symbol: "DJI",
+            from: Carbon::parse('2022-09-01'),
+            to: Carbon::parse('2022-09-05'),
+            resolution: 'D'
+        );
+
+        // Verify that the response is an object of the correct type.
+        $this->assertInstanceOf(Candles::class, $response);
+        $this->assertEmpty($response->candles);
+        $this->assertEquals($mocked_response['s'], $response->status);
+        $this->assertEquals($mocked_response['nextTime'], $response->next_time);
     }
 
     public function testExceptionHandling_throwsGuzzleException()