+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
-All notable changes to `marketdataapp-sdk-php` will be documented in this file.
+## v0.2.0-alpha
 
+- Completed Indices endpoints.
+- Added Stocks endpoints: quote, quotes, bulkQuotes, candles, bulkCandles.
+- Added custom ApiException class to handle status = 'error' messages.
+- Moved Responses to new directory.
+
+## v0.1.0-alpha
+
+- Initial release with single Indices > quote endpoint.